chore(transform): unify log record types
