+++ packages/web-sdk/src/transports/otlp/payload/transform/index.ts
@@ -1,12 +1,3 @@
-export type {
-  ErrorLogRecordPayload,
-  EventLogRecordPayload,
-  LogLogRecordPayload,
-  LogTransportItem,
-  ResourceLogPayload,
-  ResourcePayload,
-  ScopeLog,
-  LogRecordPayload,
-} from './types';
+export type { LogRecord, LogTransportItem, ResourceLog, Resource, ScopeLog } from './types';
 
 export { toScopeLog, toResourceLog } from './transform';
