+++ packages/web-sdk/src/transports/otlp/payload/transform/toErrorLogRecord.test.ts
@@ -1,16 +1,28 @@
-import { EventEvent, TransportItem, TransportItemType } from '@grafana/faro-core';
+import { ExceptionEvent, TransportItem, TransportItemType } from '@grafana/faro-core';
 
 import { toScopeLog } from './transform';
 
-const item: TransportItem<EventEvent> = {
-  type: TransportItemType.EVENT,
+const item: TransportItem<ExceptionEvent> = {
+  type: TransportItemType.EXCEPTION,
   payload: {
-    name: 'event-name',
-    domain: 'event-domain',
     timestamp: '2023-01-27T09:53:01.035Z',
-    attributes: {
-      eventAttribute1: 'one',
-      eventAttribute2: 'two',
+    type: 'Error',
+    value: 'Error message',
+    stacktrace: {
+      frames: [
+        {
+          filename: 'filename-one',
+          function: 'throwError',
+          colno: 21,
+          lineno: 11,
+        },
+        {
+          filename: 'filename-two',
+          function: 'HTMLUnknownElement.callCallback2',
+          colno: 2345,
+          lineno: 42,
+        },
+      ],
     },
     trace: {
       trace_id: 'trace-id',
@@ -48,7 +60,7 @@ const item: TransportItem<EventEvent> = {
   } as const,
 };
 
-const errorLogRecordPayload = {
+const matchErrorLogRecord = {
   timeUnixNano: 1674813181035000000,
 
   attributes: [
@@ -150,28 +162,73 @@ const errorLogRecordPayload = {
       },
     },
     {
-      key: 'event.name',
-      value: { stringValue: 'event-name' },
+      key: 'exception.type',
+      value: { stringValue: 'Error' },
     },
     {
-      key: 'event.domain',
-      value: { stringValue: 'event-domain' },
+      key: 'exception.message',
+      value: { stringValue: 'Error message' },
     },
+    // {
+    //   key: 'exception.stacktrace',
+    //   value: { stringValue: 'The unparsed stacktrace' },
+    // },
     {
-      key: 'grafana.event.attributes',
+      key: 'grafana.error.stacktrace',
       value: {
         kvlistValue: {
           values: [
             {
-              key: 'eventAttribute1',
+              key: 'frames',
               value: {
-                stringValue: 'one',
-              },
-            },
-            {
-              key: 'eventAttribute2',
-              value: {
-                stringValue: 'two',
+                arrayValue: {
+                  values: [
+                    {
+                      kvlistValue: {
+                        values: [
+                          {
+                            key: 'filename',
+                            value: { stringValue: 'filename-one' },
+                          },
+                          {
+                            key: 'function',
+                            value: { stringValue: 'throwError' },
+                          },
+                          {
+                            key: 'colno',
+                            value: { intValue: 21 },
+                          },
+                          {
+                            key: 'lineno',
+                            value: { intValue: 11 },
+                          },
+                        ],
+                      },
+                    },
+                    {
+                      kvlistValue: {
+                        values: [
+                          {
+                            key: 'filename',
+                            value: { stringValue: 'filename-two' },
+                          },
+                          {
+                            key: 'function',
+                            value: { stringValue: 'HTMLUnknownElement.callCallback2' },
+                          },
+                          {
+                            key: 'colno',
+                            value: { intValue: 2345 },
+                          },
+                          {
+                            key: 'lineno',
+                            value: { intValue: 42 },
+                          },
+                        ],
+                      },
+                    },
+                  ],
+                },
               },
             },
           ],
@@ -181,9 +238,9 @@ const errorLogRecordPayload = {
   ],
 } as const;
 
-describe('getErrorLogRecord', () => {
+describe('toErrorLogRecord', () => {
   it('Builds resource payload object for given transport item.', () => {
-    const logLogRecord = toScopeLog(item).logRecords[0];
-    expect(logLogRecord).toMatchObject(errorLogRecordPayload);
+    const errorLogRecord = toScopeLog(item).logRecords[0];
+    expect(errorLogRecord).toMatchObject(matchErrorLogRecord);
   });
 });
