+++ packages/web-sdk/src/transports/otlp/payload/types.ts
@@ -1,7 +1,7 @@
-import type { ResourceLogPayload } from './transform';
+import type { ResourceLog } from './transform';
 
 export interface OtelTransportPayload {
-  readonly resourceLogs: Readonly<ResourceLogPayload[]>;
+  readonly resourceLogs: Readonly<ResourceLog[]>;
   readonly resourceSpans: Readonly<unknown[]>;
   readonly resourceMetrics: Readonly<unknown[]>;
 }
